changed the structure of hoc
