+++ src/consts/imgSrc.ts
@@ -32,6 +32,8 @@ const imgSrc = {
   야금술심화: `${itemBaseFolder}/야금술심화.webp`,
   야금술응용: `${itemBaseFolder}/야금술응용.webp`,
   야금술특화: `${itemBaseFolder}/야금술특화.webp`,
+  야금술전문: `${itemBaseFolder}/야금술전문.webp`,
+  야금술복합: `${itemBaseFolder}/야금술복합.webp`,
   에스더의기운: `${itemBaseFolder}/esder.webp`,
   위명돌: `${itemBaseFolder}/위명돌.webp`,
   일리아칸무기: `${itemBaseFolder}/일리아칸무기.png`,
@@ -43,6 +45,8 @@ const imgSrc = {
   재봉술심화: `${itemBaseFolder}/재봉술심화.webp`,
   재봉술응용: `${itemBaseFolder}/재봉술응용.webp`,
   재봉술특화: `${itemBaseFolder}/재봉술특화.webp`,
+  재봉술전문: `${itemBaseFolder}/재봉술전문.webp`,
+  재봉술복합: `${itemBaseFolder}/재봉술복합.webp`,
   전설각인서: `${itemBaseFolder}/legendBook.webp`,
   전설각인서배경X: `${itemBaseFolder}/legendBook2.webp`,
   정제된수호강석: `${itemBaseFolder}/정제된수호강석.webp`,
