+++ src/consts/itemNameMatch.ts
@@ -1,7 +1,11 @@
 const itemNameMatch = {
   명예의파편: '명예의 파편 주머니(대)',
   야금술특화: '야금술 : 단조 특화',
+  야금술전문: '야금술 : 단조 전문',
+  야금술복합: '야금술 : 단조 복합',
   재봉술특화: '재봉술 : 수선 특화',
+  재봉술전문: '재봉술 : 수선 전문',
+  재봉술복합: '재봉술 : 수선 복합',
   야금술숙련: '야금술 : 단조 숙련',
   재봉술숙련: '재봉술 : 수선 숙련',
   파괴강석: '파괴강석',
